Split tests into categories matching documentation structure
Separate the monolithic runtests.jl into focused test files aligned
with the four major documentation categories:
- array_tests.jl: Array Operations (vecLib/vDSP)
- dsp_tests.jl: Signal Processing (FFT, DCT, convolution, filtering)
- linalg_tests.jl: Dense Linear Algebra (BLAS/LAPACK forwarding)
- sparse_tests.jl: Sparse Linear Algebra (renamed from libSparseTests.jl)
